Get module working with Backdrop
This makes a bunch of small tweaks swqpping around language to langcode
to match some of Backdrop's changes.

Additionally, there was a function that was calling element_children and
the item ID keys in this module have slashes in them, e.g. 12/en. This
throws an error in Backdrop's version of element_children so I swapped
it out with some code that matched Drupal 7's version of
element_children and it works fine.